+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/commands.py
@@ -25,7 +25,8 @@ from azure.cli.command_modules.network._client_factory import (
     cf_ddos_protection_plans, cf_public_ip_prefixes, cf_service_endpoint_policies,
     cf_service_endpoint_policy_definitions, cf_dns_references, cf_private_endpoints, cf_network_profiles,
     cf_express_route_circuit_connections, cf_express_route_gateways, cf_express_route_connections,
-    cf_express_route_ports, cf_express_route_port_locations, cf_express_route_links, cf_app_gateway_waf_policy)
+    cf_express_route_ports, cf_express_route_port_locations, cf_express_route_links, cf_app_gateway_waf_policy,
+    cf_service_tags)
 from azure.cli.command_modules.network._util import (
     list_network_resource_property, get_network_resource_property_entry, delete_network_resource_property_entry)
 from azure.cli.command_modules.network._format import (
@@ -168,7 +169,7 @@ def load_command_table(self, _):
     )
 
     network_private_endpoint_sdk = CliCommandType(
-        operations_tmpl='azure.mgmt.network.operations#InterfaceEndpointsOperations.{}',
+        operations_tmpl='azure.mgmt.network.operations#PrivateEndpointsOperations.{}',
         client_factory=cf_private_endpoints,
         min_api='2018-08-01'
     )
@@ -230,6 +231,12 @@ def load_command_table(self, _):
     network_rt_sdk = CliCommandType(
         operations_tmpl='azure.mgmt.network.operations#RouteTablesOperations.{}',
         client_factory=cf_route_tables
+    )
+
+    network_service_tags_sdk = CliCommandType(
+        operations_tmpl='azure.mgmt.network.operations#ServiceTagsOperations.{}',
+        client_factory=cf_service_tags,
+        min_api='2019-04-01'
     )
 
     network_subnet_sdk = CliCommandType(
@@ -299,10 +306,12 @@ def load_command_table(self, _):
     # endregion
 
     # region NetworkRoot
-    usage_path = 'azure.mgmt.network.operations.usages_operations#UsagesOperations.{}'
+    usage_path = 'azure.mgmt.network.operations#UsagesOperations.{}'
     with self.command_group('network') as g:
         g.command('list-usages', 'list', operations_tmpl=usage_path, client_factory=cf_usages, transform=transform_network_usage_list, table_transformer=transform_network_usage_table)
 
+    with self.command_group('network', network_service_tags_sdk) as g:
+        g.command('list-service-tags', 'list')
     # endregion
 
     # region ApplicationGateways